backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-multipart
dashscope
python-dotenv
psycopg2-binary
passlib[argon2]
pyjwt
